\name{mbgraphic}
\alias{mbgraphic}
\alias{mbgraphic}
\docType{package}
\title{
Measure Based Graphic Selection
}
\description{
Measure based exploratory data analysis. Some of the functions call interactive apps programmed with the package \pkg{shiny} to provide flexible selection options.
}
\details{
\tabular{ll}{
Package: \tab mbgraphic\cr
Type: \tab Package\cr
Version: \tab 1.0.0 \cr
Date: \tab 2017-05-07 \cr
License: \tab GPL (>= 2) \cr
}

Most important functions are: \cr
\tabular{ll}{
\code{\link[mbgraphic]{iaunivariate}} \tab Interactive app for exploring univariate anomalies \cr
\code{\link[mbgraphic]{iacorrgram}} \tab Interactive Corrgrams \cr
\code{\link[mbgraphic]{varclust}} \tab Clustering of numeric variables\cr
\code{\link[mbgraphic]{scag2sdf}} \tab Calculating (user defined) scagnostics \cr
\code{\link[mbgraphic]{sdf}} \tab Calculating measures from package \pkg{scagnostics} \cr
\code{\link[mbgraphic]{sdf_sort}} \tab Reordering a list of class \code{"sdfdata"}  \cr
\code{\link[mbgraphic]{sdf_quicksort}} \tab Fast reordering of a list of class \code{"sdfdata"} \cr
\code{\link[mbgraphic]{scaggram}} \tab Visualization of scagnostics results
in  matrix graphics \cr

\code{\link[mbgraphic]{iascaggram}} \tab Interactive scaggrams \cr
\code{\link[mbgraphic]{iascagpcp}} \tab Interactive parallel coordinate plots for exploring scagnostics results  \cr
\code{\link[mbgraphic]{selectscat}} \tab Selecting a scatterplot matrix based on scagnostics 
}
}


\author{
Katrin Grimm \cr
Department of Computer Oriented Statistics and Data Analysis \cr
University of Augsburg \cr
Germany



Maintainer: Katrin Grimm <katrin-grimm@web.de>

}

