\name{mergesdfdata}
\alias{mergesdfdata}
\title{
Merging two lists of class "sdfdata"
}
\description{
The function merges two lists of class \code{"sdfdata"} generated by function \code{\link{sdf}} or \code{\link{scag2sdf}}. 
}
\usage{
mergesdfdata(sdfdata1, sdfdata2)
}

\arguments{
  \item{sdfdata1}{
A list of class \code{"sdfdata"}.
}
  \item{sdfdata2}{
A list of class \code{"sdfdata"}.
}
}

\value{
A list of class \code{"sdfdata"}.
}

\author{
Katrin Grimm
}


\seealso{
 \code{\link{sdf}}, \code{\link{scag2sdf}}
}

\examples{
\dontrun{
data(Election2005)
set.seed(345456)
election05_small <- Election2005[,sample(5:70,5)]
scagdf <- sdf(election05_small)
addscag <- scag2sdf(election05_small,
    scagfun.list=list(dcor2d=dcor2d,splines2d=splines2d))
# merge addscag and scagsdf
scagdf2 <- mergesdfdata(scagdf,addscag)
}
}
\keyword{two-dimensional meausures and selections}